+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Pragmas and options
   when in a with-clause a projection is used in a different fixity
   (prefix vs. postfix) than in its parent clause.
 
+* New error warning `TooManyArgumentsToSort` instead of hard error.
+
 * Warning `AbsurdPatternRequiresNoRHS` has been renamed to
   `AbsurdPatternRequiresAbsentRHS`.